+++ wizard/drivers/BaseWheel.py
@@ -1,11 +1,13 @@
 #!/usr/bin/env python
 from evdev import ecodes,InputDevice, ff
 import math
+import threading
 from abc import ABC
+
 from wizard.utils.map import LinearMap
+from wizard.utils.limits import *
 from wizard.drivers.inputs import InputDevType, WheelKeyType,ControlEventType
 from wizard import config
-import threading
 
 
 class BaseWheel(ABC):
@@ -14,15 +16,14 @@ class BaseWheel(ABC):
     """
     
     # settings
-    steer_max:int = 65535 # max possible value to steering wheel
-    pedal_max:int = 255 # max possible value of pedals
+    steer_max:int = iinfo(uint16).max # max possible value to steering wheel
+    pedal_max:int = iinfo(uint8).max # max possible value of pedals
     def __init__(self, dev_type:InputDevType = InputDevType.WHEEL):
         # static variables
         self.ev_key_map:dict = {}
         self.ev_abs_map:dict = {}
         # 1 for key type and 3 for abs type
-        self.ev_events:list = [
-    ]
+        self.ev_events:list = []
         self.ev_type_accepted:tuple = (1,3)
 
         # thread
@@ -39,6 +40,7 @@ class BaseWheel(ABC):
         self.clutch_val:int = 0 # clutch input [0,255]
         # FF id
         self._ff_spring_id = None
+        self._ff_autocenter_val:int = 0
 
 
     def __del__(self):
@@ -46,12 +48,12 @@ class BaseWheel(ABC):
         self._ev.close()
 
     def _init(self):
-        # init with 49150 autocenter force
-        # self._setFF(ecodes.FF_AUTOCENTER, 49150)
+        # init with 75% autocenter force
+        self._setFFAutoCenter(49150)
         print("Racing wheel registered")
 
 
-    def SetAutoCenter(self):
+    def SetSpeedFeedback(self):
         '''
         Update the auto center force feedback using speed
         '''
@@ -63,86 +65,88 @@ class BaseWheel(ABC):
         S2W_THRESHOLD = 90
         if(speed > S2W_THRESHOLD):
             speed = S2W_THRESHOLD
-            # autocenterCmd  \in [0,65535]
-        autocenterCmd:int = int(math.sin(speed/S2W_THRESHOLD)*65535)
+        # autocenterCmd  \in [0,65535]
+        autocenterCmd:int = int((math.sin(speed/S2W_THRESHOLD)+1)/2*iinfo(uint16).max)
 
         # send autocenterCmd to the steeringwheel
-        self._setFF(ecodes.FF_AUTOCENTER, autocenterCmd)
+        self._setFFAutoCenter(autocenterCmd)
+
+    def SetWheelPos(self, val:float):
+        """
+        Set the wheel position
+        inputs:
+        - val: float to indicate wheel position [-1,1]
+        """
+        # strongest force to maintain position
+        self._setFFSpring(pos = int(val*iinfo(int16).max),
+                          saturation = iinfo(uint16).max,
+                          coeff = iinfo(int16).max)
 
 
     def start(self):
+        "start the thread"
         self._thread.start()
+
 
     def events_handler(self) -> None:
         '''
         Capture and handle events
+        TODO: set stop flag for thread
         '''
         from wizard.controller import Controller
         for event in self._ev.read_loop():
             if event.type in self.ev_type_accepted:
                 key_type:WheelKeyType = self.ev_events[event.type].get(event.code)
-                event_type:ControlEventType = self._ctl_key_map.get(key_type, ControlEventType.NONE)
+                event_type:ControlEventType = self._ctl_key_map.get(key_type,
+                                                                    ControlEventType.NONE)
                 if event_type is not ControlEventType.NONE:
                     Controller.get_instance().register_event(event_type,
-                                                    self.dev_type,event.value)
+                                                             self.dev_type,event.value)
 
 
     def _ev_connect(self):
         "Connect to evdev device based on config file"
         self._ev: evdev.InputDevice = InputDevice(config.user_input_event)
 
-    def _setFF(self,ff_type:int, val:int) -> None:
-        """
-        Set the force feedback
 
-        Input:
-        - type: type of the force feedback
-          Available Options:
-                ecodes.FF_AUTOCENTER
-                ecodes.FF_DAMPER
-                ecodes.FF_GAIN
-                ecodes.FF_FRICTION
-                ecodes.FF_SPRING
-                ecodes.FF_RAMP
-        - val: int ranging from [0,65535] or [-32768,32767]
+    def _setFFAutoCenter(self, val:int):
         """
-        if ff_type == ecodes.FF_SPRING:
-            self.setFF_spring(val)
-            return
-        assert(val >= 0 and val <= 65535)
-        self._ev.write(ecodes.EV_FF, ff_type, val)
+        Set the auto center force
+        Input: val - uint16_t indicates the intensity
+        """
+        assert(val >= 0 and val <= iinfo(uint16).max)
+        self._ff_autocenter_val = val
+        self._ev.write(ecodes.EV_FF, ecodes.FF_AUTOCENTER, val)
 
-    def setFF_spring(self, pos:int, right_saturation:int = 65535,
-                      left_saturation:int = 65535, right_coeff = 32767,
-                      left_coeff = 32767, deadband = 0) -> None:
+
+    def _setFFSpring(self, pos:int, saturation:int, coeff:int, deadband:int = 0) -> None:
         """
         Set the spring force feedback
         Input:
         - pos: position of the balance point
-        - right_saturation: maximum level when wheel moved all way to right
-        - left_saturation: same for the left side
-        - right_coeff: controls how fast the force grows when the joystick
-                       moves to the right
-        - left_coeff: same for the left side
+        - saturation: maximum level when wheel moved all way to end
+        - coeff: controls how fast the force grows when the joystick
+                 moves from center
         - deadband: size of the dead zone, where no force is produced
         """
         # create effect
         springs = (ff.Condition * 2)()
         for spring in springs:
-            spring.right_saturation = right_saturation
-            spring.left_saturation = left_saturation
-            spring.right_coeff = right_coeff
-            spring.left_coeff = left_coeff
+            spring.right_saturation = saturation
+            spring.left_saturation = saturation
+            spring.right_coeff = coeff
+            spring.left_coeff = coeff
             spring.deadband = deadband
             spring.center = pos
 
         # register effect
         spring_id = self._ev.upload_effect(
-                ff.Effect(ecodes.FF_SPRING, -1, 16384,
-                          ff.Trigger(0,0),
-                          ff.Replay(32768,0),
-                          ff.EffectType(ff_condition_effect = springs)))
+            ff.Effect(ecodes.FF_SPRING, -1, 16384,
+                      ff.Trigger(0,0),
+                      ff.Replay(iinfo(int16),0),
+                      ff.EffectType(ff_condition_effect = springs)))
 
+        # apply
         self._ev.write(ecodes.EV_FF, spring_id, 1)
 
         # erase previous effect
@@ -152,11 +156,19 @@ class BaseWheel(ABC):
         self._ff_spring_id = spring_id
 
 
-    # TODO: change
-    def erase_effect(self):
-        if self._ff_spring_id is not None:
-            self._ev.erase_effect(self._ff_spring_id)
-            self._ff_spring_id = None
+    # TODO: test if this work for all effects
+    def erase_ff(self, ff_type:int):
+        """
+        Erase the specified force feedback type
+        """
+        if ff_type == ecodes.FF_SPRING:
+            if self._ff_spring_id is not None:
+                self._ev.erase_effect(self._ff_spring_id)
+                self._ff_spring_id = None
+
+        elif ff_type == ecodes.FF_AUTOCENTER:
+            if self._ff_autocenter_val != 0:
+                self._setFFAutoCenter(0)
 
 
     @classmethod